Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
httpx>=0.27.0

# Database
sqlalchemy>=2.0.33
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# QR Code generation
qrcode[pil]>=7.4.2
pillow>=10.3.0

# JWT token signing
python-jose[cryptography]>=3.3.0

# File upload support
python-multipart>=0.0.9
